Correct grammar of documentation command
The "CatusCode" command's documentation used the indefinite article "an". This is only used when the next sound is that of a vowel. However, in this case, the next sound was /k/, which is a consonant. Therefore, the correct indefinite article should've been "a".

I have therefore changed the indefinite article from "an" to "a".