refactor(pronoun.json): Move pronoun config file
Move the pronoun configuration JSON file next to the other configuration
files, rather than being in the /utils directory. It isn't a utility;
nor does a utility directly depend on it. In addition, it's important
for it to be easy to modify. directory. It isn't a utility; nor does a
utility directly depend on it. In addition, it's important for it to be
easy to modify.